requirements.txt
django
djangorestframework
environs
psycopg2-binary
Pillow
GitPython
django-extensions